Add a comment that says it looks like lucent cards have a primary firmware revision as well (not sure which firmware versions are needed for this, but the 6.x and 8.x 'software' versions that I have seem to support it).
Add dBm comm quality RID.  This is like the normal comm quality rid,
except the signal and noise numbers are normalized to dBm.  Some
revisions of the prism firmware, however, don't support this RID, and
some that do support it return 0 for quality and/or noise.  Your
milage may vary.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@101354 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f