MFC:
a) we don't believe what the board tells us all the time (if the BIOS
hasn't run, port page 2 and port page 0 tend to be garbage)

b) add the missing code to set parameters for the SPI cards.

c) Fix copyrights in mpilib in concordance with LSI and other
freebsd developers.