Replace the vm_page hash table with a per-vmobject splay tree.  There should be no major change in performance from this change at this time but this will allow other work to progress:  Giant lock removal around VM system in favor of per-object mutexes, ranged fsyncs, more optimal COMMIT rpc's for NFS, partial filesystem syncs by the syncer, more optimal object flushing, etc.  Note that the buffer cache is already using a similar splay tree mechanism.
Note that a good chunk of the old hash table code is still in the tree.
Alan or I will remove it prior to the release if the new code does not
introduce unsolvable bugs, else we can revert more easily.

Submitted by:	alc	(this is Alan's code)
Approved by:	re


git-svn-id: svn+ssh://svn.freebsd.org/base/head@105407 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f