Change ACPI make_dev() calls to use UID_ and GID_ constants rather than hard-coded uids and gids.
Switch the device to a group of wheel instead of operator.

Narrow down the permissions on the device to require root privilege
to manipulate the system power state.  It may be that we can broaden
access to the device after review of the access control in ACPI.

Submitted by:	kris
Reviewed by:	takawata


git-svn-id: svn+ssh://svn.freebsd.org/base/head@108345 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f