Fix crash dumps on ata and scsi.
To fix scsi, don't wait for ithreads if we're dumping, it makes the
debugger sad.

To fix ata, use what appears to be a polling method if we're dumping,
I stole this from tmm but added code to ensure that this change is
only in effect while dumping.

Tested by: des


git-svn-id: svn+ssh://svn.freebsd.org/base/head@110860 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f