Remove a trailing newline, to behave as documented in syslog(3): "A trailing newline is added if none is present."
The code in syslogd, stderr, and console output always adds a newline
at the EOL.  However, the existing code never actually removed a
trailing newline, and apparently relied on syslogd to convert it
into a space character.  Thus, the existing newline was converted
to a trailing space at the EOL by syslogd, while stderr, and console
output resulted in an empty line.

MFC after:	2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@129091 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f