The change to add subversion ID has two problems. The first is that when newvers.sh is run pwd is actually the obj directory, so "../../.svn" doesn't exist and the test always fails. The second is that buildkernel is executed with a restrictive PATH, so unless you have svnversion in /bin or /usr/bin it can't run.
Fix this by looking for svnversion in /bin, /usr/bin, and /usr/local/bin
in that order. If found, store the location and derive the value of the
source directory. Then run svnversion in the appropriate directory.

There is one possible refinement which would be to add a test for
LOCALBASE!=/usr/local if we don't find svnversion the first time, but
IMO that's not necessary at this time.