Some implementations of getopt() expect that argv[0] is always the program name, and ignore that entry.  ipfw2.c code instead skips this entry and starts with options at offset 0, relying on a more tolerant implementation of the library.
This change fixes the issue by always passing a program name
in the first entry to getopt. The motivation for this change
is to remove a potential compatibility issue should we use
a different getopt() implementation in the future.

No functional changes.

Submitted by:	Marta Carbone (parts)
MFC after:	4 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@187713 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f