Change how multipath labels are created and managed. This makes it easier to support various storage boxes which really aren't active-active.
We only write the label on the *first* provider. For all other providers
we just "add" the disk. This also allows for an "add" verb.

A usage implication is that you should specificy the currently active
storage path as the first provider.

Note that this does not add RDAC-like functionality, but better allows for
autovolumefailover configurations (additional checkins elsewhere will support
this).

Sponsored by:	Panasas
MFC after:	1 month


git-svn-id: svn+ssh://svn.freebsd.org/base/head@205847 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f