sh: Add do-nothing -h option.
POSIX requires a -h option to sh and set, to locate and remember utilities
invoked by functions as they are defined. Given that this
locate-and-remember process is optional elsewhere, it seems safe enough to
make this option do nothing.

POSIX does not specify a long name for this option. Follow ksh in calling it
"trackall".


git-svn-id: svn+ssh://svn.freebsd.org/base/head@223281 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f