Japan regulatory domain update #1 - sync with CRDA entries.
Specifics:

* add 4920MHz-4980MHz for 11a and 11n/HT20
* add 5040MHz-5080MHz for 11a and 11n/HT20
* add 5500MHz-5700MHz for 11a and 11n/HT20 (DFS needed)
* add 5500MHz-5680MHz for 11n/HT40 (DFS needed)

TODO:

* add correct HT40 bands for 4920-4980 and 5040-5080

For the curious:

There's been many revisions to the Japan regulatory rules.
Apparently, the requirements require old cards certified on a previous
version of the rules to obey the older rules, not the newer rules.

The regdomain.xml outlines the -current- restrictions.

The card driver (eg the ath_hal regulatory domain code) may include
previous revisions of the Japan rules.

The ath_hal regdomain code populates the initial channel list based on
what the EEPROM indicates is possible. The regdomain.xml database imposes
further restrictions on this.

So regdomain.xml only needs to have the -current- rules. If the card
was certified on an earlier set of JP rules, it may only support a subset
of those channels - these are calculated at device attach and this
restricted list is kicked to net80211. regdomain.xml operates on -this-
list of channels.

And thus, the correct regulatory behaviour for Japan is handled with only
one regdomain.xml Japan database entry.

(phew.)

Obtained from:	Linux wireless-regdb