Introduce a new (global, sorry!) option which controls whether the ar71xx platform code should assume a uboot or redboot environment.
The current code gets very confused (and just crashes) on a uboot
environment, where each attribute=value pair is in a single entry.
Redboot on the other hand stores it as "attribute", "value", "attribute",
"value", ...

This allows the kernel to boot on a TP-LINK TL-WR1043ND from flash,
where the uboot environment gets setup. This didn't show up during a netboot
as "tftpboot" and "go" don't setup the uboot environment variables.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@227924 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f